Chá»‰nh sá»­a láº¡i data_ingestor má»™t chÃºt

+++ file_handler.py
@@ -4,15 +4,30 @@ from helper import Helper
 class FileHelper():
     def __init__(self):
         pass
+        
+    @staticmethod
+    def create_folder(folder_path:str|list[str],parrent_folder:bool=False):
+
+            
+        if folder_path:
+            if parrent_folder:
+            
+                folder_path = folder_path.rsplit("\\",1)[0]
+                os.makedirs(folder_path,exist_ok=True)
+        else:
+            print("Đường dẫn không hợp lệ")
+        
     @staticmethod
     def file_transfer(src:str,dest:str):
-        shutil.copy(src,dest)
-        
-        
-    @staticmethod
-    def create_folder(folder_path:str|list[str]):
-        os.makedirs()
-        
+        try:
+            if os.path.exists(src):
+                    shutil.copy(src,dest)
+            else:
+                Helper.show_error(None,f"Đường dẫn gốc không hợp lệ !{src}")
+       
+        except Exception as e:
+            Helper.show_error(e.errno,"File đích chưa được tạo ! hãy kiểm tra lại")
+            os._exit(0)
         
     @staticmethod    
     def remove_folder(path:str):
@@ -20,4 +35,3 @@ class FileHelper():
             shutil.rmtree(path)
         except FileNotFoundError as e : 
             Helper.show_error(e.errno,"Không tìm thấy đường dẫn file cần xóa!")
-FileHelper.remove_folder(r"C:\Users\3601183\Documents\d")
